moved Prefs to utils
